fix. int to long counter

+++ src/main/java/CancelHandler.java
@@ -1,8 +1,8 @@
 public class CancelHandler extends ActionsHandler {
-    private int transaction_id;
+    private long transaction_id;
     private int account_id;
 
-    public CancelHandler(int transaction_id, int account_id) {
+    public CancelHandler(long transaction_id, int account_id) {
         this.transaction_id = transaction_id;
         this.account_id = account_id;
     }
@@ -16,7 +16,7 @@ public class CancelHandler extends ActionsHandler {
         return account_id;
     }
 
-    public int getTransaction_id() {
+    public long getTransaction_id() {
         return transaction_id;
     }
 }
